add layer

+++ include/detectores.h
@@ -19,8 +19,8 @@
     pixel *detectorBorda(int width,int height,pixel *imagem);
     pixel *edge_gauss(long width,long height,pixel *rgb);
     pixel *median_filter(long width,long height,pixel *rgb);
-    pixel *gray_scale(long width,long height,pixel *rgb);
-    pixel *sub_image(pixel *in,pixel *out,long cols,long rows);
+    pixel *gray_scale(int width,int height,pixel *rgb);
+    pixel *sub_image(pixel *in,pixel *out,int cols,int rows);
     
 #ifdef __cplusplus
     }
